update db models

+++ backend/controllers/user.controller.js
@@ -1,6 +1,3 @@
-const bcrypt = require("bcrypt");
-const jwt = require("jsonwebtoken");
-
 const { User } = require("../models");
 
 exports.getUsers = async (req, res) => {
@@ -16,13 +13,22 @@ exports.getUsers = async (req, res) => {
 exports.deleteUser = async (req, res) => {
     try {
         const { params } = req;
+
+        if (req.auth.role !== "ADMIN") {
+            return res.status(401).json({ message: "Unauthorized" });
+        }
+
         const deleted = await User.destroy({ where: { id: params.id } });
 
         if (!deleted) {
-            return res
-                .status(404)
-                .json({ message: "User not found or not authorized" });
+            return res.status(404).json({ message: "User not found" });
         }
+
+        await Log.create({
+            action: "Delete",
+            userId: req.auth.userId,
+            table: User.tableName,
+        });
 
         res.json({ message: `User with id ${params.id} deleted` });
     } catch (err) {
